[refactor] #2 -entity, repository 패키지로 이동
