+++ src/main/java/com/Minjin/TagCafe/entity/Cafe.java
@@ -1,15 +1,11 @@
-package com.Minjin.TagCafe.entity.Cafe;
+package com.Minjin.TagCafe.entity;
 
-import com.Minjin.TagCafe.entity.CafeTag.CafeTag;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import java.time.LocalDateTime;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Table(name = "Cafe",
