+++ src/main/java/com/Minjin/TagCafe/repository/CafeRepository.java
@@ -1,7 +1,7 @@
-package com.Minjin.TagCafe.entity.Cafe;
+package com.Minjin.TagCafe.repository;
 
+import com.Minjin.TagCafe.entity.Cafe;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
