+++ src/main/java/com/Minjin/TagCafe/repository/TagRepository.java
@@ -1,5 +1,6 @@
-package com.Minjin.TagCafe.entity.Tag;
+package com.Minjin.TagCafe.repository;
 
+import com.Minjin.TagCafe.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
